Fix BrowserSync
Note sure if this is the best way to do this, but now if HTTP headers
that contain BrowserSync info will automatically be sent to the site
you're proxying. If the headers are set, then the BrowserSync script
will be added to the bottom of the page, allowing BrowserSync to work.